storage provider refactoring
